+++ Assets/Mirage/Runtime/NetworkIdentity.cs
@@ -8,7 +8,9 @@ using UnityEngine.Serialization;
 using UnityEngine.Events;
 #if UNITY_EDITOR
 using UnityEditor;
+#if UNITY_2018_3_OR_NEWER
 using UnityEditor.Experimental.SceneManagement;
+#endif
 #endif
 
 namespace Mirage
@@ -127,7 +129,7 @@ namespace Mirage
         /// This returns true if this object is the one that represents the player on the local machine.
         /// <para>This is set when the server has spawned an object for this particular client.</para>
         /// </summary>
-        public bool IsLocalPlayer => Client != null && Client.LocalPlayer == this;
+        public bool IsLocalPlayer => ClientObjectManager != null && ClientObjectManager.LocalPlayer == this;
 
         /// <summary>
         /// This returns true if this object is the authoritative player object on the client.
@@ -615,7 +617,11 @@ namespace Mirage
                     sceneId = 0;
                     // NOTE: might make sense to use GetPrefabStage for asset
                     //       path, but let's not touch it while it works.
+#if UNITY_2020_1_OR_NEWER
                     string path = PrefabStageUtility.GetCurrentPrefabStage().assetPath;
+#else
+                    string path = PrefabStageUtility.GetCurrentPrefabStage().prefabAssetPath;
+#endif
 
                     AssignAssetID(path);
                 }